Extract manual event handling logic to a ManualEventHandler
