+++ Sources/StreamChat/Workers/EventNotificationCenter.swift
@@ -17,17 +17,15 @@ class EventNotificationCenter: NotificationCenter, @unchecked Sendable {
     // Contains the ids of the new messages that are going to be added during the ongoing process
     private(set) var newMessageIds: Set<MessageId> = Set()
 
-    // The channels for which events will not be processed by the middlewares.
-    private var manualEventHandlingChannelIds: Set<ChannelId> = []
-
-    // Some events require the chat channel data, so we need to fetch it from local DB.
-    // We try to only do this once, to avoid unnecessary DB fetches.
-    private var manualEventHandlingCachedChannels: [ChannelId: ChatChannel] = [:]
+    /// Handles manual event processing for channels that opt out of middleware processing.
+    private let manualEventHandler: ManualEventHandler
 
     init(
-        database: DatabaseContainer
+        database: DatabaseContainer,
+        manualEventHandler: ManualEventHandler? = nil
     ) {
         self.database = database
+        self.manualEventHandler = manualEventHandler ?? ManualEventHandler(database: database)
         super.init()
     }
 
@@ -35,17 +33,12 @@ class EventNotificationCenter: NotificationCenter, @unchecked Sendable {
     ///
     /// The middleware's will not process events for this channel.
     func registerManualEventHandling(for cid: ChannelId) {
-        eventPostingQueue.async { [weak self] in
-            self?.manualEventHandlingChannelIds.insert(cid)
-        }
+        manualEventHandler.register(channelId: cid)
     }
 
     /// Unregister a channel for manual event handling.
     func unregisterManualEventHandling(for cid: ChannelId) {
-        eventPostingQueue.async { [weak self] in
-            self?.manualEventHandlingChannelIds.remove(cid)
-            self?.manualEventHandlingCachedChannels.removeValue(forKey: cid)
-        }
+        manualEventHandler.unregister(channelId: cid)
     }
 
     func add(middlewares: [EventMiddleware]) {
@@ -78,8 +71,7 @@ class EventNotificationCenter: NotificationCenter, @unchecked Sendable {
                     return
                 }
                 if let cid = eventDTO.payload.cid,
-                   self.manualEventHandlingChannelIds.contains(cid),
-                   let manualEvent = self.convertManualEventToDomain(event) {
+                   let manualEvent = self.manualEventHandler.handle(event) {
                     manualHandlingEvents.append(manualEvent)
                 } else {
                     middlewareEvents.append(event)
@@ -107,200 +99,6 @@ class EventNotificationCenter: NotificationCenter, @unchecked Sendable {
                 completion?()
             }
         })
-    }
-
-    private func convertManualEventToDomain(_ event: Event) -> Event? {
-        guard let eventDTO = event as? EventDTO else {
-            return nil
-        }
-
-        let eventPayload = eventDTO.payload
-
-        guard let cid = eventPayload.cid else {
-            return nil
-        }
-
-        switch eventPayload.eventType {
-        case .messageNew:
-            return createMessageNewEvent(from: eventPayload, cid: cid)
-
-        case .messageUpdated:
-            return createMessageUpdatedEvent(from: eventPayload, cid: cid)
-
-        case .messageDeleted:
-            return createMessageDeletedEvent(from: eventPayload, cid: cid)
-
-        case .reactionNew:
-            return createReactionNewEvent(from: eventPayload, cid: cid)
-
-        case .reactionUpdated:
-            return createReactionUpdatedEvent(from: eventPayload, cid: cid)
-
-        case .reactionDeleted:
-            return createReactionDeletedEvent(from: eventPayload, cid: cid)
-
-        case .channelUpdated:
-            return createChannelUpdatedEvent(from: eventPayload, cid: cid)
-
-        default:
-            return nil
-        }
-    }
-
-    // MARK: - Event Creation Helpers
-
-    private func createMessageNewEvent(from payload: EventPayload, cid: ChannelId) -> MessageNewEvent? {
-        guard
-            let userPayload = payload.user,
-            let messagePayload = payload.message,
-            let createdAt = payload.createdAt,
-            let channel = getLocalChannel(id: cid),
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else {
-            return nil
-        }
-
-        return MessageNewEvent(
-            user: userPayload.asModel(),
-            message: message,
-            channel: channel,
-            createdAt: createdAt,
-            watcherCount: payload.watcherCount,
-            unreadCount: payload.unreadCount.map {
-                .init(
-                    channels: $0.channels ?? 0,
-                    messages: $0.messages ?? 0,
-                    threads: $0.threads ?? 0
-                )
-            }
-        )
-    }
-
-    private func createMessageUpdatedEvent(from payload: EventPayload, cid: ChannelId) -> MessageUpdatedEvent? {
-        guard
-            let userPayload = payload.user,
-            let messagePayload = payload.message,
-            let createdAt = payload.createdAt,
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let channel = getLocalChannel(id: cid),
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else { return nil }
-
-        return MessageUpdatedEvent(
-            user: userPayload.asModel(),
-            channel: channel,
-            message: message,
-            createdAt: createdAt
-        )
-    }
-
-    private func createMessageDeletedEvent(from payload: EventPayload, cid: ChannelId) -> MessageDeletedEvent? {
-        guard
-            let messagePayload = payload.message,
-            let createdAt = payload.createdAt,
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let channel = getLocalChannel(id: cid),
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else { return nil }
-
-        let userPayload = payload.user
-
-        return MessageDeletedEvent(
-            user: userPayload?.asModel(),
-            channel: channel,
-            message: message,
-            createdAt: createdAt,
-            isHardDelete: payload.hardDelete
-        )
-    }
-
-    private func createReactionNewEvent(from payload: EventPayload, cid: ChannelId) -> ReactionNewEvent? {
-        guard
-            let userPayload = payload.user,
-            let messagePayload = payload.message,
-            let reactionPayload = payload.reaction,
-            let createdAt = payload.createdAt,
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let channel = getLocalChannel(id: cid),
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else { return nil }
-
-        return ReactionNewEvent(
-            user: userPayload.asModel(),
-            cid: cid,
-            message: message,
-            reaction: reactionPayload.asModel(messageId: messagePayload.id),
-            createdAt: createdAt
-        )
-    }
-
-    private func createReactionUpdatedEvent(from payload: EventPayload, cid: ChannelId) -> ReactionUpdatedEvent? {
-        guard
-            let userPayload = payload.user,
-            let messagePayload = payload.message,
-            let reactionPayload = payload.reaction,
-            let createdAt = payload.createdAt,
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let channel = getLocalChannel(id: cid),
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else { return nil }
-
-        return ReactionUpdatedEvent(
-            user: userPayload.asModel(),
-            cid: cid,
-            message: message,
-            reaction: reactionPayload.asModel(messageId: messagePayload.id),
-            createdAt: createdAt
-        )
-    }
-
-    private func createReactionDeletedEvent(from payload: EventPayload, cid: ChannelId) -> ReactionDeletedEvent? {
-        guard
-            let userPayload = payload.user,
-            let messagePayload = payload.message,
-            let reactionPayload = payload.reaction,
-            let createdAt = payload.createdAt,
-            let currentUserId = database.writableContext.currentUser?.user.id,
-            let channel = getLocalChannel(id: cid),
-            let message = messagePayload.asModel(cid: cid, currentUserId: currentUserId, channelReads: channel.reads)
-        else { return nil }
-
-        return ReactionDeletedEvent(
-            user: userPayload.asModel(),
-            cid: cid,
-            message: message,
-            reaction: reactionPayload.asModel(messageId: messagePayload.id),
-            createdAt: createdAt
-        )
-    }
-
-    private func createChannelUpdatedEvent(from payload: EventPayload, cid: ChannelId) -> ChannelUpdatedEvent? {
-        guard
-            let createdAt = payload.createdAt,
-            let channel = payload.channel?.asModel()
-        else { return nil }
-
-        let currentUserId = database.writableContext.currentUser?.user.id
-        let channelReads = channel.reads
-
-        return ChannelUpdatedEvent(
-            channel: channel,
-            user: payload.user?.asModel(),
-            message: payload.message?.asModel(cid: cid, currentUserId: currentUserId, channelReads: channelReads),
-            createdAt: createdAt
-        )
-    }
-
-    // This is only needed because some events wrongly require the channel to create them.
-    private func getLocalChannel(id: ChannelId) -> ChatChannel? {
-        if let cachedChannel = manualEventHandlingCachedChannels[id] {
-            return cachedChannel
-        }
-
-        let channel = try? database.writableContext.channel(cid: id)?.asModel()
-        manualEventHandlingCachedChannels[id] = channel
-        return channel
     }
 }
